+++ lib/src/enums/enums.dart
@@ -3,3 +3,4 @@ library;
 
 export 'authentication_type.dart';
 export 'connection_status.dart';
+export 'message_role.dart';